--- Backend/config.env
@@ -1,15 +1,0 @@
-NODE_ENV=development
-PORT=6060
-
-DATABASE = mongodb+srv://mlhgg17:<PASSWORD>@cluster0.2qg8yco.mongodb.net/MLH?retryWrites=true&w=majority
-DATABASE_PASSWORD = IJ8qpE8SA69P5coU
-
-JWT_SECRET = save-palestine-contribute
-JWT_EXPIRESIN = 1d
-
-EMAIL_USERNAME=b8ab928eb0ef33
-EMAIL_PASSWORD=4b98feb3366c2f
-EMAIL_HOST=smtp.mailtrap.io
-EMAIL_PORT=25
-
-EMAIL_FROM = aquibahmad345@gmail.com
