--- Backend/controller/authController.js
@@ -1,211 +1,0 @@
-const { promisify } = require('util');
-const jwt = require("jsonwebtoken");
-const User = require("../models/userModel");
-const catchAsync = require("../utils/catchAsync");
-const appError = require("../utils/appError");
-const crypto = require('crypto');
-const sendEmail = require('../utils/email');
-
-const sendToken = id => {
-    return jwt.sign({ id }, process.env.JWT_SECRET, {
-        expiresIn: process.env.JWT_EXPIRESIN
-    })
-}
-
-const createSendToken = (user, statusCode, res) => {
-    const token = sendToken(user._id);
-
-    res.cookie('jwt', token, {
-        expiresIn: new Date(Date.now() + process.env.JWT_COOKIE_EXPIRES_IN * 24 * 60 * 60 * 1000),
-        httpOnly: true
-    })
-
-    //remove password from output
-    user.password = undefined;
-
-    res.status(statusCode).json({
-        status: 'success',
-        token,
-        data: {
-            user
-        }
-    })
-}
-
-exports.signup = catchAsync(async (req, res) => {
-
-    const newUser = await User.create({
-        name: req.body.name,
-        email: req.body.email,
-        password: req.body.password,
-        passwordConfirm: req.body.passwordConfirm,
-    });
-
-    createSendToken(newUser, 201, res);
-});
-
-exports.login = catchAsync(async (req, res, next) => {
-
-    const { email, password } = req.body;
-    //1) check if user email and pass exist
-    if (!email || !password) {
-        return next(new appError('Please provide email and password', 400));
-    }
-    //2) check if email and password is correct
-    const user = await User.findOne({ email }).select("+password");
-
-    if (!user || !(await user.correctPassword(password, user.password))) {
-        return next(new appError('Incorrect Email or Password', 401))
-    }
-
-    //3)send jwt
-    createSendToken(user, 200, res);
-
-});
-
-exports.currentUser = catchAsync(async (req, res, next) => {
-    let token = req.cookies.jwt;
-
-    // 1) If no token move on
-    if (!token) {
-        return next();
-    }
-
-    const decoded = await promisify(jwt.verify)(token, process.env.JWT_SECRET);
-    const user = await User.findById(decoded.id);
-    if (!user) {
-        return next(new appError('The user belonging to this token does no longer exist.', 401))
-    }
-    res.status(200).json({
-        status: 'success',
-        data: {
-            user
-        }
-    })
-})
-
-exports.protect = catchAsync(async (req, res, next) => {
-    //1) Getting token and check if it's there
-    let token;
-    if (req.headers.authorization && req.headers.authorization.startsWith('Bearer')) {
-        token = req.headers.authorization.split(' ')[1]
-    } else if (req.cookies.jwt) {
-        token = req.cookies.jwt;
-    }
-
-    if (!token) {
-        return next(new appError('You are not logged in! Please log in to get access.', 401))
-    }
-    //2) Verification token
-    const decoded = await promisify(jwt.verify)(token, process.env.JWT_SECRET)
-    //3) Check if user still exists
-    const currentUser = await User.findById(decoded.id)
-    if (!currentUser) {
-        return next(new appError('The user belonging to this token does no longer exist.', 401))
-    }
-    //4) Check if user changed password after the token was issued
-    if (currentUser.changedPasswordAfter(decoded.iat)) {
-        return next(new appError('User recently changed password! Please log in again.', 401))
-    }
-    //GRANT ACCESS TO PROTECTED ROUTE
-    req.user = currentUser;
-    next();
-});
-
-exports.isLoggedIn = async (req, res, next) => {
-    if (req.cookies.jwt) {
-        try {
-            // 1) verify token
-            const decoded = await promisify(jwt.verify)(
-                req.cookies.jwt,
-                process.env.JWT_SECRET
-            );
-
-            // 2) Check if user still exists
-            const currentUser = await User.findById(decoded.id);
-            if (!currentUser) {
-                return next();
-            }
-
-            // 3) Check if user changed password after the token was issued
-            if (currentUser.changedPasswordAfter(decoded.iat)) {
-                return next();
-            }
-
-            // THERE IS A LOGGED IN USER
-            res.locals.user = currentUser;
-
-            console.log(currentUser);
-            return next();
-        } catch (err) {
-            return next();
-        }
-    }
-    next();
-};
-exports.forgotPassword = catchAsync(async (req, res, next) => {
-    // 1) Get user based on POSTed email
-    const user = await User.findOne({ email: req.body.email });
-    if (!user) {
-        return next(new appError('There is no user with email address.', 404));
-    }
-
-    // 2) Generate the random reset token
-    const resetToken = user.createPasswordResetToken();
-    await user.save({ validateBeforeSave: false });
-
-    // 3) Send it to user's email
-    const resetURL = `${req.protocol}://${req.get('host')}/api/v1/users/resetPassword/${resetToken}`;
-
-    const message = `Forgot your password? Submit a PATCH request with your new password and passwordConfirm to: ${resetURL}.\nIf you didn't forget your password, please ignore this email!`;
-
-    try {
-        await sendEmail({
-            email: user.email,
-            subject: 'your password reset token (valid for 10 min)',
-            message
-        })
-
-        res.status(200).json({
-            status: 'success',
-            message: 'Token sent to Email'
-        })
-    } catch (err) {
-        user.passwordResetToken = undefined;
-        user.passwordResetExpires = undefined;
-        await user.save({ validateBeforeSave: false });
-
-        return next(
-            new appError('There was an error sending the email. Try again later!'),
-            500
-        );
-    }
-
-    next();
-})
-
-exports.resetPassword = catchAsync(async (req, res, next) => {
-    // 1) Hash the token in the url 
-    const hashedToken = crypto
-        .createHash('sha256')
-        .update(req.params.token)
-        .digest('hex');
-    //Get the user based on the hashed token
-    const user = await User.findOne({ passwordResetToken: hashedToken, passwordResetExpires: { $gt: Date.now() } });
-
-    // 2) If the user doesnot exist send error
-    if (!user) {
-        return next(new appError('Token is invalid or has expired', 400));
-    }
-    user.password = req.body.password;
-    user.passwordConfirm = req.body.passwordConfirm;
-    user.passwordResetToken = undefined;
-    user.passwordResetExpires = undefined;
-
-    await user.save();
-    // 3) update changedPasswordAt property 
-
-    // 4) Log the user in, send JWT
-    createSendToken(user, 20, res);
-    next();
-})
