--- Backend/package.json
@@ -1,27 +1,0 @@
-{
-  "dependencies": {
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "express-mongo-sanitize": "^2.2.0",
-    "express-rate-limit": "^7.1.2",
-    "helmet": "^7.0.0",
-    "hpp": "^0.2.3",
-    "jsonwebtoken": "^9.0.2",
-    "mongodb": "^6.2.0",
-    "mongoose": "^7.6.3",
-    "nodejs": "^0.0.0",
-    "nodemon": "^3.0.1",
-    "parcel-bundler": "^1.12.5",
-    "xss-clean": "^0.1.4",
-    "zaproxy": "^2.0.0-rc.3"
-  },
-  "name": "mlh",
-  "version": "1.0.0",
-  "description": "vulnerability scanner",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "SKM",
-  "license": "ISC"
-}
