--- Backend/script.py
@@ -1,40 +1,0 @@
-#!/usr/bin/env python
-import time
-from pprint import pprint
-from zapv2 import ZAPv2
-import sys
-
-apiKey = 'ps78svv227qmk7ctdi77dcjf6m';
-target = sys.argv[1];
-
-# By default ZAP API client will connect to port 8080
-zap = ZAPv2(apikey=apiKey)
-# Use the line below if ZAP is not listening on port 8080, for example, if listening on port 8090
-# zap = ZAPv2(apikey=apiKey, proxies={'http': 'http://127.0.0.1:8090', 'https': 'http://127.0.0.1:8090'})
-
-print('Spidering target {}'.format(target))
-# The scan returns a scan id to support concurrent scanning
-scanID = zap.spider.scan(target)
-
-while int(zap.spider.status(scanID)) < 100:
-    # Poll the status until it completes
-    print('Spider progress %: {}'.format(zap.spider.status(scanID)))
-    time.sleep(1)
-
-print('Spider has completed!')
-# Prints the URLs the spider has crawled
-print('\n'.join(map(str, zap.spider.results(scanID))))
-# If required post process the spider results
-
-
-while int(zap.pscan.records_to_scan) > 0:
-    # Loop until the passive scan has finished
-    print('Records to passive scan : ' + zap.pscan.records_to_scan)
-    time.sleep(2)
-
-print('Passive Scan completed')
-
-# Print Passive scan results/alerts
-print('Hosts: {}'.format(', '.join(zap.core.hosts)))
-print('Alerts: ')
-pprint(zap.core.alerts())
