--- Backend/utils/appError.js
@@ -1,13 +1,0 @@
-class appError extends Error {
-    constructor(message, statusCode) {
-        super(message);
-
-        this.statusCode = statusCode;
-        this.status = `${statusCode}`.startsWith('4') ? 'fail' : 'error';
-        this.isOperational = true;
-
-        Error.captureStackTrace(this, this.constructor);
-    }
-}
-
-module.exports = appError;
