--- Backend/utils/catchAsync.js
@@ -1,5 +1,0 @@
-module.exports = fn => {
-    return (req, res, next) => {
-        fn(req, res, next).catch(next);
-    };
-};
