--- Backend/views/_footer.pug
@@ -1,51 +1,0 @@
-.footer-section#Contact
-  .container
-    footer.py-5.my-4
-      ul#footer-list-item.nav.justify-content-center.border-bottom.pb-3.mb-3
-        li.nav-item
-          ul.footer-contact-list.px-5
-            li#contact-list-item.footer-list
-              th
-                h4 Resources
-            li#contact-list-item.footer-list
-              //- td: a(href="", class="") img(style="height: 50px;", src="", alt="") What's New
-            li#contact-list-item.footer-list
-              td: a(href="", class="")  Privacy Ploicy
-        li.nav-item
-          ul.footer-contact-list.px-5
-            li#contact-list-item.footer-list
-              th
-                h4 Get Started
-            li#contact-list-item.footer-list
-              td: a#footer-list-content(href="/signup") Sign up now
-        li.nav-item
-          ul.footer-contact-list.px-5
-            li#contact-list-item.footer-list
-              th
-                h4 About Us
-            li#contact-list-item.footer-list
-              td: a#footer-list-content(href="#") Our Team
-        li.nav-item
-          ul.footer-contact-list.px-5
-            li#contact-list-item.footer-list
-              th
-                h4 Contact Us
-            li#contact-list-item.footer-list
-              td
-                div.location
-                  i.fa-solid.fa-location-dot
-                  |  Jamia Hamdard,
-                  br
-                  |  New Delhi - 110065
-      //- Social Icon
-      .container.footer-icon
-        a.social-icon(href="#")
-          i.fa-brands.fa-instagram
-        a.social-icon(href="#")
-          i.fa-brands.fa-x-twitter
-        a.social-icon(href="#")
-          i.fa-brands.fa-linkedin-in
-        a.social-icon(href="mailto: ")
-          i.fa-solid.fa-envelope
-      p.footer-copyright.text-center.text-body-secondary
-        | &copy; GLITCH HACKERS 2023
