--- Backend/views/base.pug
@@ -1,40 +1,0 @@
-doctype html
-html(lang="en")
-  head
-    meta(charset="UTF-8")
-    meta(name="viewport", content="width=device-width, initial-scale=1.0")
-    title VUL SCANNER
-    // Bootstrap CSS Scripts
-    link(href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css", rel="stylesheet", integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65", crossorigin="anonymous")
-    
-    // CSS Stylesheet
-    link(rel="stylesheet", href="css/style.css")
-    script(src="script.js")
-
-    // Font Awesome
-    link(rel="stylesheet", href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.1/font/bootstrap-icons.css")
-    script(src="https://kit.fontawesome.com/1c582dc9cf.js", crossorigin="anonymous")
-    
-    // type
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/typed.js/2.0.11/typed.min.js")
-    //- script(src="https://cdnjs.cloudflare.com/ajax/libs/OwlCarousel2/2.3.4/owl.carousel.min.js")
-
-    // Google Fonts
-    link(rel="preconnect", href="https://fonts.googleapis.com")
-    link(rel="preconnect", href="https://fonts.gstatic.com", crossorigin)
-    link(href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Ubuntu:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap", rel="stylesheet")
-
-  body  
-    
-    include _header
-
-    block content
-
-    include _footer 
-
-    script(type='module' src="/authGetStart.js")
-    script(type='module' src="/login.js")
-    script(type='module' src="/scan.js")
-    script(type='module' src="/script.js")
-    script(src="https://cdnjs.cloudflare.com/ajax/libs/axios/1.6.0/axios.min.js")
-    script(src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js", integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4", crossorigin="anonymous")
