+++ script.js
@@ -1,0 +1,7 @@
+ // typing text animation script
+ var typed = new Typed(".typing", {
+    strings: ["Defend, Detect, and Stay Secure Online"],
+    typeSpeed: 150,
+    // backSpeed: 60,
+    loop: true
+});
