+++ style.css
@@ -1,0 +1,435 @@
+/* Google Fonts - Poppins */
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap');
+
+*{
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Poppins', sans-serif;
+}
+h1,
+h2,
+h3,
+h4 {
+    font-family: 'poppins', sans-serif;
+}
+
+h1 {
+    font-size: 3.5rem;
+}
+
+h2 {
+    font-size: 2rem;
+}
+
+p {
+    font-size: 1rem;
+    letter-spacing: 0.04rem;
+}
+
+a {
+    letter-spacing: 0.02rem;
+}
+body{
+    background-color: #fff;
+}
+/* Navbar */
+.navbar {
+    position: absolute;
+    top: 0;
+    width: 100%;
+    padding: 0.4rem;
+    background-color: #333333;
+    z-index: 1000;
+}
+
+.navbar-brand {
+    padding-left: 6rem;
+    font-size: 2rem;
+    font-weight: bold;
+}
+
+.navbar-color-section {
+    background-color: #333333;
+}
+
+.nav-item {
+    padding: 0.6rem;
+}
+
+.offcanvas-body{
+    padding-left: -20%;
+}
+.nav-button {
+    padding: 15px;
+    border-radius: 30%;
+}
+
+.nav-link {
+    font-size: 1.2rem;
+}
+
+
+/* OffCanvas section */
+.offcanvas-title {
+    color: #ffffff;
+    font-weight: 600;
+}
+
+/* index page */
+.main{
+    background: rgba(0, 0, 0, 1) url("pic1.jpg") no-repeat center;
+    background-size: cover;
+    height: 100vh;
+    /* margin-top: 150px; */
+}
+.home{
+    display: flex;
+    font-family: 'Ubuntu', sans-serif;
+    /* margin-top: 20%; */
+    padding-top: 20%;
+    margin-left: 23%;
+}
+.home .max-width{
+  width: 100%;
+  display: flex;
+}
+.home .max-width .row{
+  margin-right: 0;
+}
+
+.home .home-content .text-3{
+    font-size: 40px;
+    margin: 5px 0;
+}
+.home .home-content .text-3 span{
+    color: #00cc00;
+    font-weight: 500;
+}
+
+/* get started */
+.signUp{
+    color: #fff;
+    width: 30%;
+    margin-top: 5%;
+    margin-left: 30%;
+}
+
+/* about section styling */
+.about{
+    padding-top: 1.5%;
+    background-color: #333333;
+    color: #FFF;
+}
+
+.about .about-content .left{
+    width: 45%;
+    padding-left: 15%;
+    display: inline;
+    vertical-align: middle; 
+    text-align: right;
+}
+.about .about-content .left img{
+    height: 190px;
+    width: 290px;
+    object-fit: cover;
+    margin-top: -40px;
+}
+.about .about-content .right{
+    margin-left: 5%;
+    display: inline-block;
+    width: 50%;
+}
+.about .about-content .right p{
+    width: 85%;
+    text-align: justify;
+}
+
+/* our team */
+.team{
+    background-color: #333333;
+    color: #fff;
+}
+.card{
+    color: #333333;
+}
+.row{
+    padding-left: 18%;
+    padding-bottom: 5%;
+}
+
+/* output */
+.output, .prof{
+    padding-top: 10%;
+}
+
+/* profile page */
+.prof{
+    padding-top: 10%;
+    margin-left: -10%;
+}
+
+
+/* Login page */
+/* form */
+.containerA{
+    height: 100vh;
+    width: 100%;
+    margin-top: 5%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: #c3c3c4;
+    column-gap: 30px;
+}
+.form{
+    position: absolute;
+    max-width: 430px;
+    width: 100%;
+    padding: 30px;
+    border-radius: 6px;
+    background: #FFF;
+}
+.form.signup{
+    opacity: 0;
+    pointer-events: none;
+}
+.forms.show-signup .form.signup{
+    opacity: 1;
+    pointer-events: auto;
+}
+.forms.show-signup .form.login{
+    opacity: 0;
+    pointer-events: none;
+}
+header{
+    font-size: 28px;
+    font-weight: 600;
+    color: #232836;
+    text-align: center;
+}
+
+.form .field{
+    position: relative;
+    height: 50px;
+    width: 100%;
+    margin-top: 20px;
+    border-radius: 6px;
+}
+.field input,
+.field button{
+    height: 100%;
+    width: 100%;
+    border: none;
+    font-size: 16px;
+    font-weight: 400;
+    border-radius: 6px;
+}
+.field input{
+    outline: none;
+    padding: 0 15px;
+    border: 1px solid#CACACA;
+}
+.field input:focus{
+    border-bottom-width: 2px;
+}
+.eye-icon{
+    position: absolute;
+    top: 50%;
+    right: 10px;
+    transform: translateY(-50%);
+    font-size: 18px;
+    color: #8b8b8b;
+    cursor: pointer;
+    padding: 5px;
+}
+.field button{
+    color: #fff;
+    background-color: #D1E0ED;
+    transition: all 0.3s ease;
+    cursor: pointer;
+}
+.form-link{
+    text-align: center;
+    margin-top: 10px;
+}
+.form-link span, 
+.form-link a{
+    font-size: 14px;
+    font-weight: 400;
+    color: #232836;
+}
+.form-check{
+    font-size: 14px;
+}
+.form a{
+    color: #0171d3;
+    text-decoration: none;
+}
+.form-content a:hover{
+    text-decoration: underline;
+}
+.line{
+    position: relative;
+    height: 1px;
+    width: 100%;
+    margin: 36px 0;
+    background-color: #d4d4d4;
+}
+.line::before{
+    content: 'Or';
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: #FFF;
+    color: #8b8b8b;
+    padding: 0 15px;
+}
+.media-options a{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+/* btn */
+.login-buttons {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+  }
+  
+  .login-buttons button {
+    background: none;
+    border: none;
+    border-radius: 50%;
+    padding: 5px;
+    cursor: pointer;
+    font-size: 24px;
+    color: #333; 
+    transition: color 0.3s;
+  }
+  
+  .login-buttons button:hover {
+    color: #7e8287;   
+  }
+
+
+/* home or post login */
+.input-group{
+    padding-top: 20%;
+    width: 50%;
+    margin-left: 23%;
+}
+
+
+
+
+
+
+
+
+
+/* Footer */
+.footer-section {
+    background-color: #000000;
+    position: absolute;
+    width: 100%;
+}
+
+.footer-icon {
+    display: flex;
+    justify-content: center;
+}
+
+.social-icon {
+    color: #ffffff;
+    font-size: x-large;
+    margin: 20px 10px;
+    padding: 0 1rem 0;
+}
+
+.footer-list {
+    font-size: large;
+    font-weight: 600;
+    list-style-type: none;
+    color: #ffffff;
+    margin: 10px 10px;
+}
+
+.footer-contact-list {
+    margin: 0;
+    padding: 0;
+}
+
+#contact-list-item {
+    margin-left: 0;
+    margin-right: 0;
+    letter-spacing: 0.02rem;
+    font-size: medium;
+    font-weight: 400;
+}
+
+li>a {
+    color: #ffffff;
+    font-size: medium;
+    font-weight: 400;
+    text-decoration: none;
+    cursor: pointer;
+
+}
+
+.footer-copyright {
+    color: #ffffff;
+    padding: 15px 0;
+}
+
+/* Responsive footer section */
+@media (max-width: 1200px) {
+    #footer-list-item {
+        display: grid;
+        grid-template-columns: repeat(2, 1fr);
+        text-align: start;
+    }
+
+    @media (max-width: 500px) {
+        #footer-list-item {
+            display: grid;
+            grid-template-columns: repeat(1, 1fr);
+            text-align: start;
+        }
+    }
+}
+@media screen and (max-width: 400px) {
+    .form{
+        padding: 20px 10px;
+    }
+}
+
+
+/* index page */
+@media (max-width: 500px) {
+.home{
+    display: flex;
+    font-family: 'Ubuntu', sans-serif;
+    margin-top: 15%;
+    margin-left: 15%; 
+}
+.main{
+    height: 80vh;
+}
+/* home page */
+.input-group{
+    padding-top: 40%;
+    width: 80%;
+    margin-left: 13%;
+}
+.navbar{
+    width: fit-content;
+}
+    .navbar-brand {
+        padding-left: 0%;
+        font-size: 2rem;
+        font-weight: bold;
+    }
+}
